+++ pkg/quarkus/v1alpha/scaffolds/internal/templates/doc.go
