+++ pkg/quarkus/v1alpha/scaffolds/internal/templates/gitignore.go
