+++ pkg/quarkus/v1alpha/scaffolds/internal/templates/model/model.go
@@ -5,7 +5,7 @@ import (
 
 	"sigs.k8s.io/kubebuilder/v3/pkg/machinery"
 
-	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1/scaffolds/internal/templates/util"
+	"github.com/operator-framework/java-operator-plugins/pkg/quarkus/v1alpha/scaffolds/internal/templates/util"
 )
 
 var _ machinery.Template = &Model{}
