+++ pkg/quarkus/v1alpha/scaffolds/internal/templates/pomxml.go
