+++ pkg/quarkus/v1alpha/scaffolds/internal/templates/util/file.go
