+++ pkg/quarkus/v1alpha/util/util.go
